Safari-609.1.11.1/LayoutTests/webrtc/video-setDirection-expected.txt



PASS Going from sendrecv to inactive and back to sendrecv 
FAIL The MediaStream should remain the same assert_equals: expected object "[object MediaStream]" but got object "[object MediaStream]"
